docker_mote/requirements.txt
discord.py
redis
nltk